+++ changelog.md
@@ -1,3 +1,8 @@
+# Version 2024.8.4 (2024-08-16)
+
+- Make load only relevant paramset descriptions
+- Add UN_IGNORE_WILDCARD to get_parameters
+
 # Version 2024.8.3 (2024-08-15)
 
 - Ignore parameters on un ignore parameter list